requirements.txt
absl-py==2.3.0
alembic==1.16.1
astunparse==1.6.3
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
dash==3.0.4
flake8==7.2.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.58.2
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.73.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mypy==1.16.0
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.42.1
nest-asyncio==1.6.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
optuna==4.3.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
protobuf==4.25.8
pvlib==0.13.0
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
retrying==1.3.4
rich==14.0.0
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.0.6
wheel==0.45.1
wrapt==1.17.2
zipp==3.23.0
